backend/requirements.txt
mysql-connector-python
Flask==2.3.3
Werkzeug==2.3.7
flask-cors